chore: move formula cut test to sheets-formula-ui package
